requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
gunicorn==23.0.0
kombu==5.6.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
unicorn==2.1.4
vine==5.1.0
wcwidth==0.2.14
